Moved core datastructure related puzzles to datastructures-algorithms project
