+++ src/main/java/puzzle/io/PrintPyramid.java
@@ -1,7 +1,5 @@
-package quiz.io;
+package puzzle.io;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Scanner;
 
 public class PrintPyramid {
@@ -22,19 +20,4 @@ public class PrintPyramid {
             System.out.print(sb.toString());
         }
     }
-
-
-
-    public ArrayList<String> replaceEmployee(ArrayList<String> employee, String replaceThis, final String replaceWith)
-    {
-        ArrayList<String> employeeList = new ArrayList<>();
-        for (String name : employee) {
-            if (name.equals(replaceThis))
-                employeeList.add(replaceWith);
-            else
-                employeeList.add(name);
-        }
-        return employeeList;
-    }
-
 }
